Started general level layout and design, imported starting points from old projects
I added in a map i drew sketching out a general layout of a potential test level. I then traced over it with probuilder. I plan to use this same workflow on the main levels, except I will then export the probuilder mesh and make it more interesting and consistent in blender. I also added in the general movement system, as well as a test character to stand in until I finish the mech model and animations. This character also had the rebindable controls, settings, and pause menus as dependencies so I imported them all in as well. I updated the menus and scripts to only be relevant to this jam.